Make it work, dirtily

+++ libft/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: lgasc <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/05 17:55:31 by lgasc             #+#    #+#             */
-/*   Updated: 2023/10/24 12:03:56 by lgasc            ###   ########.fr       */
+/*   Updated: 2024/01/10 18:37:09 by lgasc            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 ///@returns The substring. NULL if the allocation fails.
 ///@remarks External function: `malloc`
 ///@see https://stackoverflow.com/a/5416724
-char	*ft_substr(char const *superstring, unsigned int start, size_t length)
+char	*ft_substr(char const *superstring, size_t start, size_t length)
 {
 	char			*substring;
 	unsigned int	i;
